CassetteNote/backend_fastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
PyJWT
passlib
python-dotenv
pydantic
pydantic-settings
aiofiles
Pillow
bcrypt
